1-task/requirements.txt
torch==2.0.1
torchaudio==2.0.2
librosa==0.10.0
pytorch-lightning==2.0.2
pandas==2.0.3
numpy==1.24.3
tqdm==4.65.0
scikit-learn==1.2.2
torchmetrics==0.11.4
PyYAML==6.0 